Add files via upload
activity_main.xml acts as the view for the login page. It is currently not very stylized however that is low priority for now and can be done later when we have decided a color scheme and have necessary code done. This file is used as the view with MainActivity.kt. MainActivity.kt provides some control for the login screen. It has unfinished code inside of the onclick listeners for the log in and new user button. Strings.xml is also included.